LoginView 분리, ChatVC 기본 세팅
